WIP
After a couple of agonizing hours debugging a missing needsUpdate
property, I finally have a procedurally generated displacement map. The
results are disappointing. There are now visible seams in the meshes
and, although the shadows cast on the surface of the meshes by the point
light look reasonably good, I did not realize that the shadows cast on
the backdrop and neighboring meshes would remain spherical.

I also just learned through this workshop about Three’s ability to
export meshes as OBJs. Given my desire for better shadows and the
opportunity to render with Blender, I’m going back to deforming the
meshes.